StateOfCharge Enhancer considers coulombic efficiency
Now, the StateOfCharge RawDataEnhancer takes into account Coulombic Efficiency,
and outputs a third column to the raw dataframe, that of the CE adjusted charge,
meaning, the actual charge in the battery, rather than the one observed by the
cycler. It defaults to 1.0 (100% efficiency), so, by default, this new column
is equivalent to the `'cycled_charge'` column.