Postprocess modiffications (#145)
* ignoring vscode settings

* changing name of schema from hold to rest

* Fixing nomenclature errors

Changing the ChargingState schema from 'hold' to 'rest' caused a number of
issues, which should now have been fixed.

* Changing the way we assign basic control methods

The way to assign the simpler control methods is more streamlined. Short rests
are tied only to their duration, independently of how many data points are
present. Other short steps are classified as either pulses if they have >=5 data
points, or short non-rests, in case there are <5 data points. Non-rest steps
that last longer than 30 seconds but have fewer than 5 datapoints are considered
unknown control method.

* style fix

* Allowing user to specify max duration of short period

When computing the control method, we now allow the user to specify what they
wish to consider is a short enough period to be classified as short rest, pulse,
or short non-rest. That value defaults to 30 seconds, as originally intended.

* Additional explanation to ControlMethod schema

Included additional explanation for some of the pre-specified control methods.

* StateOfCharge Enhancer considers coulombic efficiency

Now, the StateOfCharge RawDataEnhancer takes into account Coulombic Efficiency,
and outputs a third column to the raw dataframe, that of the CE adjusted charge,
meaning, the actual charge in the battery, rather than the one observed by the
cycler. It defaults to 1.0 (100% efficiency), so, by default, this new column
is equivalent to the `'cycled_charge'` column.

* style fix

* fixing demo notebook