Postprocess CE fix (#147)
* Avoiding warnings

The way pandas deals with the addition of new columns is a bit weird, as there
is a difference between a view of a dataframe, a copy of it, and the dataframe
itself. By using `.loc` with a specified tuple of new columns to add, we do what
pandas deems correct, and circumvent several annoying warnings

* Changing the name of the CE column

The CE-adjusted charge column has a new, more descriptive name.

* trying to avoid the nuisance of Coverall 422

* trying again to deal with coveralls